Fix gps high baudrates

+++ board/project/nmea.h
@@ -40,7 +40,7 @@
 
 typedef struct nmea_parameters_t
 {
-    uint16_t baudrate;
+    uint32_t baudrate;
     float *lat, *lon, *alt, *spd, *cog, *hdop, *sat, *time, *date, *vspeed, *dist, *spd_kmh;
 } nmea_parameters_t;
